requirements.txt
google-genai>=1.0.0
pyautogui>=0.9.54
pillow>=10.0.0
python-dotenv>=1.0.1
mss>=9.0.1
pytesseract>=0.3.10
pydantic>=2.0
pyperclip>=1.8.0
websocket-client>=1.6.0
# Run replay dashboard (optional — only required when running
# `python -m agent --serve-dashboard`). The agent itself never imports
# FastAPI; the dashboard module lazy-imports it.
fastapi>=0.110.0
uvicorn>=0.27.0
httpx>=0.27.0  # used by FastAPI's TestClient in the dashboard tests
